Split Key class into PrivateKey and PublicKey
